Logout button
Basic HTTP Authentication is an ancient(15+ -year-old) auth method that
wasn't designed to manage logging out. Thus the only way to log out is
closing the browser(not only the current tab). There are some
hacks/workaround online but they are not comprehensively compatible to
different browsers or OS.

For now I am just adding a pop up script to reminds users to close the
browser if they want to log out or don't want their submission to be
messed by evil roommates.

If in the future we want a "real" logout functionality, we have to
switch to a session-aware authentication protocol.

**This is also related to Issue #76.
link: https://github.com/RCOS-Grading-Server/HWserver/issues/76